Extract `GhostItemCardView` to its own file with preview.
